+++ R/rcpp_roxy.R
@@ -1,0 +1,3 @@
+#' @useDynLib multinets
+#' @importFrom Rcpp sourceCpp
+NULL