+++ tests/testthat/test-mode_transformation.R
@@ -5,7 +5,12 @@ test_that("mode transformation responds properly to different objects", {
   data(g.bip, package = "sand")
   data(rfid, package = "igraphdata")
 
-  expect_error(mode_transformation(linked_sim))
-  expect_error(mode_transformation(rfid))
-  expect_silent(mode_transformation(igraph::upgrade_graph(g.bip)))
+  expect_error(mode_transformation(linked_sim, "high"))
+  expect_error(mode_transformation(rfid, "high"))
+  expect_silent(mode_transformation(igraph::upgrade_graph(g.bip), "high"))
+})
+
+test_that("function with missing argument 'which' raises error", {
+  data(rfid, package = "igraphdata")
+  expect_error(mode_transformation(igraph::upgrade_graph(g.bip)))
 })