+++ NEWS.md
@@ -1,3 +1,12 @@
+# multinets 0.1.1
+
+* *bug fix* - issue #1 - fix if statement with condition length > 1. Changed chain of if statements in mode_transformation to a switch and included an error raise when which argument is missing.
+
+* *enhance* set_shape_multilevel performance with cpp code.
+
+* include Rcpp in the project.
+
+
 # multinets 0.1.0
 
 * First release